+++ Changelog.md
@@ -5,6 +5,7 @@ OpenCore Changelog
 - Documentation improvements (see Differences.pdf)
 - Platform information database updates
 - Fixed misbehaving Debug -> Target enable bit
+- Added `ResetLogoStatus` ACPI quirk
 
 #### v0.0.1
 - Initial developer preview release