+++ Changelog.md
@@ -6,6 +6,8 @@ OpenCore Changelog
 - Platform information database updates
 - Fixed misbehaving Debug -> Target enable bit
 - Added `ResetLogoStatus` ACPI quirk
+- Added `SpoofVendor` PlatformInfo feature
+- Replaced `ExposeBootPath` with `ExposeSensitiveData`
 
 #### v0.0.1
 - Initial developer preview release