+++ Changelog.md
@@ -4,10 +4,13 @@ OpenCore Changelog
 #### v0.0.2
 - Documentation improvements (see Differences.pdf)
 - Platform information database updates
-- Fixed misbehaving Debug -> Target enable bit
+- Fixed misbehaving `Debug` -> `Target` enable bit
 - Added `ResetLogoStatus` ACPI quirk
 - Added `SpoofVendor` PlatformInfo feature
 - Replaced `ExposeBootPath` with `ExposeSensitiveData`
+- Added builtin implimentation of DataHub protocol
+- Dropped `UpdateSMBIOSMode` `Auto` mode in favour of `Create`
+- Fixed SMBIOS CPU detection for Xeon and Core models
 
 #### v0.0.1
 - Initial developer preview release