+++ Changelog.md
@@ -13,6 +13,7 @@ OpenCore Changelog
 - Fixed SMBIOS CPU detection for Xeon and Core models
 - Moved `ConsoleControl` configuration to `Protocols`
 - Added `Security` -> `ScanPolicy` preference
+- Fixed invalid `board-rev` exposure in Data Hub
 
 #### v0.0.1
 - Initial developer preview release